--- app/models/BasicPokemonDetail.js
@@ -1,190 +1,0 @@
-'use strict';
-
-var mongoose = require('mongoose');
-
-var BasicPokemonDetailSchema = new mongoose.Schema({
-    pokemonID: {
-        type: Number,
-        unique: true,
-        index: true
-    },
-    name: {
-        type: String
-    },
-    gender: {
-        abbreviation: {
-            type: String, enum: ['m', 'f', 'g', 'h'] //m=male, f=female, g=genderless, h=hermaphrodite
-        },
-        male_ratio: Number,
-        female_ratio: Number,
-        breedable: Boolean
-    },
-    height: {
-        value: Number
-    },
-    weight: {
-        value: Number
-    },
-    category: {
-        bug: Boolean,
-        grass: Boolean,
-        dark: Boolean,
-        ground: Boolean,
-        dragon: Boolean,
-        ice: Boolean,
-        electric: Boolean,
-        normal: Boolean,
-        fairy: Boolean,
-        poison: Boolean,
-        fighting: Boolean,
-        psychic: Boolean,
-        fire: Boolean,
-        rock: Boolean,
-        flying: Boolean,
-        steel: Boolean,
-        ghost: Boolean,
-        water: Boolean
-    },
-    move: {
-        fury_cutter: Boolean,
-        bug_bite: Boolean,
-        bite: Boolean,
-        sucker_punch: Boolean,
-        dragon_breath: Boolean,
-        thunder_shock: Boolean,
-        spark: Boolean,
-        low_kick: Boolean,
-        karate_chop: Boolean,
-        ember: Boolean,
-        wing_attack: Boolean,
-        peck: Boolean,
-        lick: Boolean,
-        shadow_claw: Boolean,
-        vine_whip: Boolean,
-        razor_leaf: Boolean,
-        mud_shot: Boolean,
-        ice_shared: Boolean,
-        frost_breath: Boolean,
-        quick_attack: Boolean,
-        scratch: Boolean,
-        tackle: Boolean,
-        pound: Boolean,
-        cut: Boolean,
-        poison_jab: Boolean,
-        acid: Boolean,
-        psycho_cut: Boolean,
-        rock_throw: Boolean,
-        metal_claw: Boolean,
-        bullet_punch: Boolean,
-        water_gun: Boolean,
-        splash: Boolean,
-        mud_slap: Boolean,
-        zen_headbutt: Boolean,
-        confusion: Boolean,
-        poison_sting: Boolean,
-        bubble: Boolean,
-        feint_attack: Boolean,
-        steel_wing: Boolean,
-        fire_fang: Boolean,
-        rock_smash: Boolean,
-        wrap: Boolean,
-        hyper_beam: Boolean,
-        dark_pulse: Boolean,
-        sludge: Boolean,
-        vice_grip: Boolean,
-        flame_wheel: Boolean,
-        megahorn: Boolean,
-        flamethrower: Boolean,
-        dig: Boolean,
-        cross_chop: Boolean,
-        psybeam: Boolean,
-        earthquake: Boolean,
-        stone_edge: Boolean,
-        ice_punch: Boolean,
-        heart_stamp: Boolean,
-        discharge: Boolean,
-        flash_cannon: Boolean,
-        drill_peck: Boolean,
-        ice_beam: Boolean,
-        blizzard: Boolean,
-        heat_wave: Boolean,
-        aerial_ace: Boolean,
-        drill_run: Boolean,
-        petal_blizzard: Boolean,
-        mega_drain: Boolean,
-        bug_buzz: Boolean,
-        poison_fang: Boolean,
-        night_slash: Boolean,
-        bubble_beam: Boolean,
-        submission: Boolean,
-        low_sweep: Boolean,
-        aqua_jet: Boolean,
-        aqua_tail: Boolean,
-        seed_bomb: Boolean,
-        psyshock: Boolean,
-        ancient_power: Boolean,
-        rock_tomb: Boolean,
-        rock_slide: Boolean,
-        power_gem: Boolean,
-        shadow_sneak: Boolean,
-        shadow_punch: Boolean,
-        ominous_wind: Boolean,
-        shadow_ball: Boolean,
-        magnet_bomb: Boolean,
-        iron_head: Boolean,
-        parabolic_charge: Boolean,
-        thunder_punch: Boolean,
-        thunder: Boolean,
-        thunderbolt: Boolean,
-        twister: Boolean,
-        dragon_pulse: Boolean,
-        dragon_claw: Boolean,
-        disarming_voice: Boolean,
-        draining_kiss: Boolean,
-        dazzling_gleam: Boolean,
-        moonblast: Boolean,
-        play_rough: Boolean,
-        cross_poison: Boolean,
-        sludge_bomb: Boolean,
-        sludge_wave: Boolean,
-        gunk_shot: Boolean,
-        bone_club: Boolean,
-        bulldoze: Boolean,
-        mud_bomb: Boolean,
-        signal_beam: Boolean,
-        x_scissor: Boolean,
-        flame_charge: Boolean,
-        flame_burst: Boolean,
-        fire_blast: Boolean,
-        brine: Boolean,
-        water_pulse: Boolean,
-        scald: Boolean,
-        hydro_pump: Boolean,
-        psychic: Boolean,
-        psystrike: Boolean,
-        icy_wind: Boolean,
-        giga_drain: Boolean,
-        fire_punch: Boolean,
-        solar_beam: Boolean,
-        leaf_blade: Boolean,
-        power_whip: Boolean,
-        air_cutter: Boolean,
-        hurricane: Boolean,
-        brick_break: Boolean,
-        swift: Boolean,
-        horn_attack: Boolean,
-        stomp: Boolean,
-        hyper_fang: Boolean,
-        body_slam: Boolean,
-        rest: Boolean,
-        struggle: Boolean
-    },
-    evolution: {
-        type: [String]
-    },
-    pre_form: {
-        type: [String]
-    }
-});
-
-module.exports = mongoose.model('BasicPokemonDetail', BasicPokemonDetailSchema);
