+++ startUp.js
@@ -1,7 +1,6 @@
 (function () {
     "use strict";
     let express = require('express'),
-        loadBasicPokemonDetails = require('./app/controllers/loadBasicPokemonDetails'),
         bodyParser = require('body-parser'),
         cors = require('cors'),
         config = require('./config');
@@ -57,9 +56,6 @@
             app.listen(port);
             logger.info('Listening on port ' + port);
 
-
-            //load basic pokemon details
-            loadBasicPokemonDetails.start();
         },
         stop: function (done) {
             logger.info('Application stopped');
